move wallet to components folder
